Update the save_scalar() method to accept and store the timestamp at along with the scalar value (#170)
* Adding support to save the scalars with user provided timestamp.

* Removed the unnecessary global variables

* Re-organized the save scalar tests as per frameworks

* Addressed the review comment.

* Addressed the review comments

* Adding timestamp for write_sumary methods
