+++ smdebug/core/writer.py
@@ -119,7 +119,9 @@ class FileWriter:
         smd = SummaryMetadata(plugin_data=plugin_data)
         return smd
 
-    def write_tensor(self, tdata, tname, write_index=True, mode=ModeKeys.GLOBAL, mode_step=None):
+    def write_tensor(
+        self, tdata, tname, write_index=True, mode=ModeKeys.GLOBAL, mode_step=None, timestamp=None
+    ):
         mode, mode_step = self._check_mode_step(mode, mode_step, self.step)
         smd = self._get_metadata(mode, mode_step)
         value = make_numpy_array(tdata)
@@ -127,9 +129,11 @@ class FileWriter:
         tensor_proto = make_tensor_proto(nparray_data=value, tag=tag)
         s = Summary(value=[Summary.Value(tag=tag, metadata=smd, tensor=tensor_proto)])
         if write_index:
-            self._writer.write_summary_with_index(s, self.step, tname, mode, mode_step)
+            self._writer.write_summary_with_index(
+                s, self.step, tname, mode, mode_step, timestamp=timestamp
+            )
         else:
-            self._writer.write_summary(s, self.step)
+            self._writer.write_summary(s, self.step, timestamp)
 
     def write_graph(self, graph):
         self._writer.write_graph(graph)
@@ -145,8 +149,8 @@ class FileWriter:
         event = Event(tagged_run_metadata=trm)
         self._writer.write_event(event)
 
-    def write_summary(self, summ, global_step):
-        self._writer.write_summary(summ, global_step)
+    def write_summary(self, summ, global_step, timestamp: float = None):
+        self._writer.write_summary(summ, global_step, timestamp=timestamp)
 
     def write_histogram_summary(self, tdata, tname, global_step, bins="default"):
         """Add histogram data to the event file.
@@ -179,9 +183,9 @@ class FileWriter:
         except ValueError as e:
             logger.warning(f"Unable to write histogram {tname} at {global_step}: {e}")
 
-    def write_scalar_summary(self, name, value, global_step):
+    def write_scalar_summary(self, name, value, global_step, timestamp: float = None):
         s = scalar_summary(name, value)
-        self._writer.write_summary(s, global_step)
+        self._writer.write_summary(s, global_step, timestamp=timestamp)
 
     def flush(self):
         """Flushes the event file to disk.
